refactor(github-analysis): split AllContributorsCommitChart into reusable components
- Extract chart-related logic and components into separate files
- Create new components: CommitLineChart, ChartTip, ChartTooltipContent, NoDataDisplay
- Move chart data preparation and utilities to chart-utils.ts
- Update index.ts to export new components and types
- Refactor AllContributorsCommitChart to use new components and utilities
