+++ src/components/github/analysis/components/index.ts
@@ -8,4 +8,11 @@ export * from './controls';
 export * from './containers';
 
 // Stats components
-export * from './stats'; 
+export * from './stats';
+
+export { default as AllContributorsCommitChart } from './AllContributorsCommitChart';
+export { default as NoDataDisplay } from './NoDataDisplay';
+export { default as ChartTip } from './ChartTip';
+export { default as CommitLineChart } from './CommitLineChart';
+export { default as ChartTooltipContent } from './ChartTooltipContent';
+export * from './chart-types';
